+++ tests/test_data_structures.py
@@ -1,5 +1,5 @@
-import os
 import pathlib
+import shutil
 import tempfile
 import unittest
 from unittest.mock import patch
@@ -52,30 +52,45 @@ class DataStructuresTestCase(unittest.TestCase):
 
 
 class DockerArgsTest(unittest.TestCase):
+    goldens_dir_path = pathlib.Path(__file__).parent.joinpath("goldens")
+
+    def setUp(self) -> None:
+        self.build_temp_dir = pathlib.Path(tempfile.mkdtemp())
+        dags_path = pathlib.Path(self.build_temp_dir).joinpath("dag")
+        dags_path.mkdir(parents=True)
+        shutil.copytree(
+            self.goldens_dir_path.joinpath("config"), dags_path.joinpath("config")
+        )
+
+    def tearDown(self) -> None:
+        shutil.rmtree(self.build_temp_dir)
+
     @patch("data_pipelines_cli.data_structures.git_revision_hash")
     def test_build_tag(self, mock_git_revision_hash):
-        repository = "rep"
+        repository = "my_docker_repository_uri"
         commit_sha = "eee440bfbe0801ec3f533f897c1d55e6a5afd5cd"
         mock_git_revision_hash.return_value = commit_sha
 
-        docker_args = DockerArgs(repository)
+        with patch("data_pipelines_cli.cli_constants.BUILD_DIR", self.build_temp_dir):
+            docker_args = DockerArgs("base")
+
         self.assertEqual(f"{repository}:{commit_sha}", docker_args.docker_build_tag())
         self.assertEqual(repository, docker_args.repository)
         self.assertEqual(commit_sha, docker_args.commit_sha)
+
+    @patch("data_pipelines_cli.cli_constants.BUILD_DIR", goldens_dir_path)
+    @patch("data_pipelines_cli.data_structures.git_revision_hash")
+    def test_no_repository(self, mock_git_revision_hash):
+        commit_sha = "eee440bfbe0801ec3f533f897c1d55e6a5afd5cd"
+        mock_git_revision_hash.return_value = commit_sha
+
+        with self.assertRaises(DataPipelinesError):
+            _ = DockerArgs("base")
 
     @patch("data_pipelines_cli.data_structures.git_revision_hash")
     def test_no_git_hash(self, mock_git_revision_hash):
         mock_git_revision_hash.return_value = None
 
-        with self.assertRaises(DataPipelinesError):
-            _ = DockerArgs("rep")
-
-    @patch("data_pipelines_cli.data_structures.git_revision_hash")
-    @patch.dict(os.environ, {"REPOSITORY_URL": "repo"})
-    def test_no_repository_argument(self, mock_git_revision_hash):
-        commit_sha = "eee440bfbe0801ec3f533f897c1d55e6a5afd5cd"
-        mock_git_revision_hash.return_value = commit_sha
-
-        docker_args = DockerArgs(None)
-        self.assertEqual("repo", docker_args.repository)
-        self.assertEqual(commit_sha, docker_args.commit_sha)
+        with patch("data_pipelines_cli.cli_constants.BUILD_DIR", self.build_temp_dir):
+            with self.assertRaises(DataPipelinesError):
+                _ = DockerArgs("base")